pyproject.toml
[tool.poetry]
name = "mobile_tests_appium"
version = "0.1.0"
description = "mobile end-to-end testing on browserstack with selene + appium + pytest"
authors = ["qaguru <qaguru@gmail.com>"]
packages = [{include = "mobile_tests_appium"}]

[tool.poetry.dependencies]
python = "^3.10"
selene = {version = "^2.0.0-beta.8", allow-prereleases = true}
Appium-Python-Client = "^2.6.1"
pytest = "^7.1.3"
pytest-xdist = "^2.5.0"
allure-pytest = "^2.10.0"
requests = "^2.28.1"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
target-version = ['py310']
skip-string-normalization = true


[tool.pytest.ini_options]
addopts = "--alluredir=allure-results"